Backend: move client to kube/client
The client here is for talking to the Kubernetes backend (and doing port
forwarding with it); that is entirely Kuberentes-specific and not used in
the case of a setup with Kubernetes disabled.  It is clearer to move it out
of the way so we can put clients dealing with the container engine in its
place.
