+++ pkg/rancher-desktop/backend/kube/lima.ts
@@ -17,8 +17,8 @@ import INSTALL_K3S_SCRIPT from '@pkg/assets/scripts/install-k3s';
 import LOGROTATE_K3S_SCRIPT from '@pkg/assets/scripts/logrotate-k3s';
 import SERVICE_CRI_DOCKERD_SCRIPT from '@pkg/assets/scripts/service-cri-dockerd.initd';
 import SERVICE_K3S_SCRIPT from '@pkg/assets/scripts/service-k3s.initd';
-import { KubeClient } from '@pkg/backend/client';
 import * as K8s from '@pkg/backend/k8s';
+import { KubeClient } from '@pkg/backend/kube/client';
 import { ContainerEngine } from '@pkg/config/settings';
 import mainEvents from '@pkg/main/mainEvents';
 import { checkConnectivity } from '@pkg/main/networking';
