+++ pkg/rancher-desktop/backend/kube/wsl.ts
@@ -5,9 +5,9 @@ import util from 'util';
 
 import semver from 'semver';
 
-import { KubeClient } from '../client';
 import K3sHelper, { ExtraRequiresReasons, NoCachedK3sVersionsError, ShortVersion } from '../k3sHelper';
 import WSLBackend, { Action } from '../wsl';
+import { KubeClient } from './client';
 
 import INSTALL_K3S_SCRIPT from '@pkg/assets/scripts/install-k3s';
 import { BackendSettings, RestartReasons } from '@pkg/backend/backend';
